Automating the flow for auto triggering training and retraining in CI CD pipeline (#7)
* CI/CD code changes

* CI/CD code changes

* CI/CD code changes

* Update DATASET_VERSION

* Update DATASET_VERSION

* CI/CD code changes

* CI/CD code changes

* CI/CD code changes

* CI/CD code changes

* CI/CD code changes

* Update DATASET_VERSION

* CI/CD code changes

* Updated CI.yml with check changes

* CI/CD code changes

* CI/CD code changes

* Update CI.yml with changes of clearml

* Update CI.yml with updated pipeline name

* docker file changes

* Update DATASET_VERSION

* update the re-train in CI

* Update DATASET_VERSION testing

* adding wrapper of training

* Update DATASET_VERSION check for retraining

* Update CI.yml with new runner-"CI-runner"

* Update CI.yml

* Update DATASET_VERSION

* Update CI.yml

* Update DATASET_VERSION

* Update DATASET_VERSION

* Update DATASET_VERSION

* Update DATASET_VERSION

* Update CI.yml

* Update CI.yml

* Update CI.yml

* Update DATASET_VERSION

* updated keras folder

* Update CI.yml

* Update DATASET_VERSION

* correcting task variable useage

* making training by default

* Add new file to repository

* Add new file to repository

* Add new file to repository

* Adding upload_model script

* adding h5 file

* Update DATASET_VERSION

* fixing issue with upload model code

* data manager remove_old_pipeline code fix

* fixing issues

* trying git save

* adding git config in CI pipeline

* adding git config in CI pipeline

* adding git config in CI pipeline

* testing git push

* commiting keras file

* commiting keras file

* commiting keras file

* commiting keras file

* commiting keras file

* git test file

* fixing git issues using subprocess

* revising code for fixing git issues

* updated file

* updated ci ymal for emailid

* udated

* updated

* updated

* commiting keras and h5 files to git

* retraining

* retrain

* commiting keras and h5 files to git

* retrain with 2 epochs

* commiting keras and h5 files to git

* Stabling CICD pieline (#6)

* Update README.md added codespace info

* Update README.md

* Added FastApi code in main.py for prometheus and changed Dockerfile and requirement.txt for same

* Added fastapi code in main.py for prometheus .dockerfile and rwuirement.txt for same

* Update README.md

---------


* Update CI.yml

---------

+++ harit_model/processing/data_manager.py
@@ -25,9 +25,14 @@ def save_pipeline(pipeline_to_persist) -> None:
 def remove_old_pipelines(*, files_to_keep: t.List[str]) -> None:
     """Remove old model pipelines."""
     do_not_delete = files_to_keep + ["__init__.py"]
-    for model_file in TRAINED_MODEL_DIR.iterdir():
-        if model_file.name not in do_not_delete:
-            model_file.unlink()
+    if(os.path.exists(TRAINED_MODEL_DIR)):
+        print ("iterating inside directory",TRAINED_MODEL_DIR )
+        for model_file in TRAINED_MODEL_DIR.iterdir():
+            if model_file.name not in do_not_delete:
+                model_file.unlink()
+    else: 
+        TRAINED_MODEL_DIR.mkdir(parents=True, exist_ok=True)
+        print ("created directory ",TRAINED_MODEL_DIR)
 
 def copy_folder(source_folder, destination_folder):
     """Copy all files and subdirectories from source to destination."""
